requirements.txt
streamlit==1.28.1
python-docx==0.8.11
pdfplumber==0.10.3
google-generativeai==0.3.0
python-dotenv==1.0.0
pillow==10.0.0
